fix warning

+++ include/lock_free_queue.hpp
@@ -34,7 +34,7 @@ public:
         for (size_t i = 0; i < _capacity; ++i)
         {
             _queue[i].tail.store(i, std::memory_order_relaxed);
-            _queue[i].head.store(-1, std::memory_order_relaxed);
+            _queue[i].head.store((size_t)(-1), std::memory_order_relaxed);
         }
 
         _tail.store(0, std::memory_order_relaxed);
